requirements.txt
python-telegram-bot==20.6
APScheduler==3.10.1
requests
beautifulsoup4
